requirements.txt
pillow>=10.0.0
keyring>=24.0.0
python-dotenv>=1.0.0
openai>=1.0.0
psutil>=5.9.0
